requirements.txt
python-telegram-bot==11.1.0
vk-api==11.9.0
dialogflow==0.7.2
google-api-core==1.23.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0